[BUGFIX] deleteRecord_flexFormCallBack deletes file_references
Files that are referenced in a flexform
(<type>group</type><internal_type>file_reference</internal_type>)
will be deleted by t3lib_TCEmain::deleteRecord_flexFormCallBack(),
independent of the path or usage by other records.

For normal record fields (e.g. tt_content.images), only file with the
internal_type "file" are deleted. Files of internal_type "file_reference"
are respected and will not been deleted. There is an appropriate condition
in t3lib_TCEmain::extFileFunctions().

But for flexform fields, this differentiation in handling between both
internal_types is missing.

Therefore using lowlevel_cleaner or the delete option in the recycler
will delete files that may be still in use. The flexFormCallBack must
be changed to delete only files of internal_type="file".

Releases: 4.7, 4.6, 4.5
Fixes: #40052

+++ t3lib/class.t3lib_tcemain.php
@@ -4125,8 +4125,8 @@ class t3lib_TCEmain {
 		$refIndexObj = t3lib_div::makeInstance('t3lib_refindex');
 		$files = $refIndexObj->getRelations_procFiles($dataValue, $dsArr['TCEforms']['config'], $PA['uid']);
 
-			// Traverse files and delete them:
-		if (is_array($files)) {
+			// Traverse files and delete them if the field is a regular file field (and not a file_reference field)
+		if (is_array($files) && $dsArr['TCEforms']['config']['internal_type'] == 'file') {
 			foreach ($files as $dat) {
 				if (@is_file($dat['ID_absFile'])) {
 					unlink($dat['ID_absFile']);
